requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
black==22.12.0
    # via -r requirements.in
click==8.1.3
    # via black
flake8==6.0.0
    # via -r requirements.in
isort==5.11.2
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mypy==0.991
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
pathspec==0.10.3
    # via black
platformdirs==2.6.2
    # via black
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
python-frontmatter==1.0.0
    # via -r requirements.in
pyyaml==6.0
    # via python-frontmatter
tomli==2.0.1
    # via
    #   black
    #   mypy
typing-extensions==4.4.0
    # via mypy
